Title: 십진수와 이진수, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 231117/십진수와 이진수/decimal-and-binary-number.cpp
@@ -6,7 +6,7 @@ using namespace std;
 int main() {
     string s;
     cin >> s;
-    int n = 0, c = 1;
+    long long n = 0, c = 1;
     for(int i=s.length()-1;i>=0;i--) {
         n += (int)(s[i] - 48) * c;
         c *= 2;
